Move dbSNP query to ncbi module
This removes 25% of the code from the Retriever module which didn't
really belong there in the first place.

No output object is used anymore, a simple exception is thrown on
connection errors. We also removed the auto retry around the Entrez call
which I think didn't fix a problem that's ours to fix and added a lot of
complexity.